+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog of checkoutmanager
 1.10 (unreleased)
 -----------------
 
+- Copying sample .cfg file if it doesn't exist instead of only suggesting the
+  copy. Fixes #12.
+
 - Added hidden info command.  Should be only useful for subversion if
   your svn program is updated and your OS requires you to give svn
   access to your stored credentials again, for each repository.